BookEvent moved to Parser
