+++ src/coinbase_pro/parser.hpp
@@ -6,7 +6,6 @@
 #pragma once
 #include "book_l2.hpp"
 #include <common/parser.hpp>
-#include <functional>
 #include <rapidjson/document.h>
 #include <string_view>
 #include <unordered_map>
@@ -15,7 +14,6 @@ namespace ssc2ce {
 
 class CoinbaseParser : public Parser {
 public:
-  using BookEvent = std::function<void(BookL2 *)>;
   CoinbaseParser();
   ~CoinbaseParser() {}
 
@@ -31,23 +29,10 @@ public:
     last_error_msg_.clear();
   }
 
-  BookL2 const *get_book(const std::string_view &instrument);
-
-  void set_on_book_setup(BookEvent handler)
-  {
-    on_book_setup_ = handler;
-  }
-
-  void set_on_book_update(BookEvent handler)
-  {
-    on_book_update_ = handler;
-  }
+  BookL2 const *get_book(const std::string_view &instrument) override;
 
 private:
   std::string last_error_msg_;
-
-  BookEvent on_book_setup_;
-  BookEvent on_book_update_;
 
   std::unordered_map<std::size_t, CoinbaseBookL2> books_;
   CoinbaseBookL2 &find_or_create_book(const std::string_view &instrument);
